requirements.txt
# Medical Prescription Extraction System Requirements

# Core dependencies
python-dotenv==1.0.0           # Environment variable management
numpy==1.24.3                  # Numerical operations
scipy==1.11.2                  # Scientific computing

# Audio processing
librosa==0.10.0                # Audio loading and processing
soundfile==0.12.1              # WAV/FLAC file I/O
pydub==0.25.1                  # Audio manipulation

# Speech recognition
openai==1.3.0                  # OpenAI API (Whisper)
groq==0.4.1                    # Groq API
faster-whisper==0.10.0         # Local Whisper inference

# NLP and text processing
langdetect==1.0.9              # Language detection
textblob==0.17.1               # Text processing
nltk==3.8.1                    # Natural Language Toolkit
regex==2023.10.3               # Advanced regex support

# Database
sqlite3                         # Built-in SQLite support (Python 3.9+)

# Web and API
httpx==0.24.1                  # HTTP client
requests==2.31.0               # HTTP library

# Data processing and serialization
pandas==2.0.3                  # Data manipulation
pydantic==2.3.0                # Data validation

# Machine learning (optional, for smart labeling)
scikit-learn==1.3.1            # ML utilities
torch==2.0.1                   # PyTorch (for ML models, optional)
transformers==4.33.2           # Hugging Face transformers (optional)

# Logging and monitoring
python-json-logger==2.0.7      # JSON logging

# Testing
pytest==7.4.2                  # Testing framework
pytest-cov==4.1.0              # Coverage reporting
pytest-asyncio==0.21.1         # Async test support

# Code quality
black==23.10.0                 # Code formatter
flake8==6.1.0                  # Linter
pylint==3.0.2                  # Code analyzer
isort==5.12.0                  # Import sorting

# Development
ipython==8.15.0                # Interactive shell
jupyter==1.0.0                 # Jupyter notebooks
notebook==7.0.4                # Jupyter Notebook UI

# Optional: Web framework for API
fastapi==0.103.1               # Modern web framework
uvicorn==0.23.2                # ASGI server
pydantic-settings==2.0.3       # Settings management

# Optional: Database ORM
sqlalchemy==2.0.22             # SQL toolkit and ORM

# Performance
tqdm==4.66.1                   # Progress bars
